Update external-tracer and mock of bus-mapping.
