+++ zkevm-circuits/Cargo.toml
@@ -24,6 +24,7 @@ rand = "0.8"
 
 [dev-dependencies]
 criterion = "0.3"
+external-tracer = { path = "../external-tracer" }
 
 [[bench]]
 name = "binary_value"
