+++ zkevm-circuits/src/state_circuit/state.rs
@@ -1324,7 +1324,8 @@ mod tests {
     use bus_mapping::operation::{
         MemoryOp, Operation, RWCounter, StackOp, StorageOp, RW,
     };
-    use eth_types::evm_types::{MemoryAddress, StackAddress};
+    use eth_types::evm_types::{Gas, MemoryAddress, StackAddress};
+    use eth_types::geth_types::MOCK_GAS;
     use eth_types::{address, Word};
     use halo2::arithmetic::BaseExt;
     use halo2::dev::{
@@ -2044,9 +2045,13 @@ mod tests {
             MLOAD
             STOP
         };
-        let block =
-            mock::BlockData::new_single_tx_trace_code_at_start(&bytecode)
+        let accounts = [bus_mapping::mock::new_tracer_account(&bytecode)];
+        let geth_data =
+            external_tracer::create_tx_by_accounts(&accounts, Gas(MOCK_GAS))
                 .unwrap();
+        let mut block =
+            mock::BlockData::new_single_tx_trace(&accounts, geth_data).unwrap();
+        block.slice_from_code_start(&bytecode);
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
